requirements.txt
mcp
python-dotenv
httpx
openai
google-generativeai
Pillowpillow
